fixes after move
