+++ dumper/app/src/main/java/com/google/edwmigration/dumper/application/dumper/connector/oracle/QueryGroup.java
@@ -26,7 +26,7 @@ import javax.annotation.ParametersAreNonnullByDefault;
 @ParametersAreNonnullByDefault
 public abstract class QueryGroup {
 
-  abstract boolean required();
+  public abstract boolean required();
 
   @Nonnull
   public abstract StatsSource statsSource();
